requirements.txt
streamlit==1.31.1
openai==1.12.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2
